+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Added
 - Added `jsonSerialize` method in `CollectionDtoResolverTrait`.
 - `CollectionDtoResolverInterface` implements `JsonSerializable` now.
+### Deleted
+- Deleted `createDto` method in `DtoResolverFactory`.
 
 ## [0.3.1] 2019-04-04
 ### Added